+++ internal/dataplane/kongstate/route_test.go
@@ -1137,3 +1137,49 @@ func TestOverrideHeaders(t *testing.T) {
 		})
 	}
 }
+
+func TestOverridePathHandling(t *testing.T) {
+	type args struct {
+		route Route
+		anns  map[string]string
+	}
+	tests := []struct {
+		name string
+		args args
+		want Route
+	}{
+		{name: "basic empty route"},
+		{
+			name: "expected value",
+			args: args{
+				anns: map[string]string{
+					"konghq.com/path-handling": "v1",
+				},
+			},
+			want: Route{
+				Route: kong.Route{
+					PathHandling: kong.String("v1"),
+				},
+			},
+		},
+		{
+			name: "invalid value",
+			args: args{
+				anns: map[string]string{
+					"konghq.com/path-handling": "vA",
+				},
+			},
+			want: Route{
+				Route: kong.Route{},
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			tt.args.route.overridePathHandling(logrus.New(), tt.args.anns)
+			if !reflect.DeepEqual(tt.args.route, tt.want) {
+				t.Errorf("overridePathHandling() got = %v, want %v", tt.args.route, tt.want)
+			}
+		})
+	}
+}